refactor(desktop): 抽离 usePrNavigation，App 卸下 PR 导航/范围领域
把发现分类、活跃·归档范围切换、归档冷存储懒加载、按 URL 打开、定位跳转（jumpToPr）、
系统通知点击导航，以及跨组件 Diff/Tab 跳转意图（pendingDiffNav / pendingTab）收进
hooks/usePrNavigation，组合在 usePullRequests（列表 + 选中）之上。App 仅保留组合根
装配与廉价视图模型派生。App.tsx 474 → 320 行，零行为变更。
